Clean up for real
