+++ benchmarks/internal/bench.js
