+++ benchmarks/internal/choose-worker.js
