+++ benchmarks/internal/cluster/worker.js
@@ -1,5 +1,5 @@
 'use strict'
-const { ClusterWorker } = require('../../lib/index')
+const { ClusterWorker } = require('../../../lib/index')
 
 function yourFunction (data) {
   for (let i = 0; i <= 1000; i++) {
